Scale identification
This file allows one to identify which scale is played in an array of notes. I will need to modify it for the case where there is not a single scale played, like in most jazz standards (but this is more complicated, especially when the players starts playing outside the key).